requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aniso8601=8.0.0=py_0
attrs=19.3.0=py_0
ca-certificates=2019.11.28=hecc5488_0
certifi=2019.11.28=py36_0
click=7.0=py_0
flask=1.1.1=py_0
flask-restplus=0.13.0=py36_0
importlib_metadata=1.0.0=py36_0
itsdangerous=1.1.0=py_0
jinja2=2.10.3=py_0
jsonschema=3.2.0=py36_0
libcxx=9.0.0=h89e68fa_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h6de7cb9_1006
markupsafe=1.1.1=py36h0b31af3_0
more-itertools=7.2.0=py_0
ncurses=6.1=h0a44026_1002
openssl=1.1.1d=h0b31af3_0
pip=19.3.1=py36_0
pyrsistent=0.15.6=py36h0b31af3_0
python=3.6.9=h359304d_0
python-dateutil=2.8.1=py_0
pytz=2019.3=py_0
readline=7.0=hcfe32e1_1001
setuptools=42.0.1=py36_0
six=1.13.0=py36_0
sqlite=3.30.1=ha441bb4_0
tk=8.6.10=hbbe82c9_0
werkzeug=0.16.0=py_0
wheel=0.33.6=py36_0
xz=5.2.4=h1de35cc_1001
zipp=0.6.0=py_0
zlib=1.2.11=h0b31af3_1006
